hack-a-league-main/requirements.txt
streamlit
opencv-python
mediapipe
pyautogui
pytesseract
pillow
numpy
openai
python-dotenv
plotly
